getInitSettings
